bit of ecstasy cleanup

+++ src/Speed/GameCube/Src/Ecstasy/EcstasyE.hpp
@@ -7,8 +7,8 @@
 
 #include "Speed/GameCube/Src/Ecstasy/eViewPlat.hpp"
 
+// total size: 0x40
 struct eRenderTarget {
-    // total size: 0x40
     TARGET_ID ID;             // offset 0x0, size 0x4
     const char *Name;         // offset 0x4, size 0x4
     int Active;               // offset 0x8, size 0x4
@@ -16,7 +16,7 @@ struct eRenderTarget {
     int ScissorY;             // offset 0x10, size 0x4
     int ScissorW;             // offset 0x14, size 0x4
     int ScissorH;             // offset 0x18, size 0x4
-    int FrameAddress;         // offset 0x1C, size 0x4
+    intptr_t FrameAddress;    // offset 0x1C, size 0x4
     int FrameWidth;           // offset 0x20, size 0x4
     int FrameHeight;          // offset 0x24, size 0x4
     FILTER_ID CopyFilterID;   // offset 0x28, size 0x4
